// Milestone2_FrontEnd/StockMarketFrontEnd/src/app/Model/company.ts
export class COMPANY {
    id: number;
    company_code: string;
    company_name: string;
    turnover: number;
    ceo: string;
    board_of_director: string;
    listed_in_se: string;
    sector_name: string;
    brife_write_up: string;
    stock_code: string;
    company_status: string;
}
// for Manage Company
export class COMPANY1 {
    company_code: string;
    companyName: string;
    turnover: number;
    ceo: string;
    boardofdirectors: string;
    listedinse: string;
    sectorName: string;
    brifewriteup: string;
    stockCode: string;
    companyStatus: string;
}